+++ packages/spacecat-shared-ticket-client/CHANGELOG.md
@@ -1,3 +1,9 @@
+## [@adobe/spacecat-shared-ticket-client-v1.0.4](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-ticket-client-v1.0.3...@adobe/spacecat-shared-ticket-client-v1.0.4) (2026-07-13)
+
+### Bug Fixes
+
+* **ticket-client:** fix token refresh logic, listIssueTypes endpoint, and createTicket validation  ([#1801](https://github.com/adobe/spacecat-shared/issues/1801)) ([cfe80d8](https://github.com/adobe/spacecat-shared/commit/cfe80d8fae220c9d76b6ac9f8cbd5df4ab363e3c))
+
 ## [@adobe/spacecat-shared-ticket-client-v1.0.3](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-ticket-client-v1.0.2...@adobe/spacecat-shared-ticket-client-v1.0.3) (2026-07-08)
 
 ### Bug Fixes